fix(musea): support strict port for serve
